third_party/blink/web_tests/external/wpt/css/css-scroll-snap/capturing-snap-positions-expected.txt

This is a testharness.js-based test.
[FAIL] The third item should be snapped to by default, not the second's child.
  assert_equals: expected 200 but got 100
Harness: the test ran to completion.
